refactor : 훅 분리 및 파일명수정, 테스트 코드 수정:
